pyproject.toml
[project]
name = "pyiss"
version = "0.3.0"
description = "OPUS-backed helper tools for Cassini ISS"
requires-python = ">=3.10"
dependencies = ["requests", "pandas", "numpy", "pillow"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
